feat: redis and bug

+++ src/controllers/roleController.js
@@ -1,14 +1,15 @@
+/* eslint-disable no-undef */
 const supabase = require('../config/supabaseClient');
 
 const roleController = {
   // Menambahkan role baru
   addRole: async (request, h) => {
-    const { roleWorker } = request.payload;
+    const { roleworker } = request.payload;
 
     try {
       const { error } = await supabase
         .from('role_worker')
-        .insert([{ roleWorker }]);
+        .insert([{ roleworker }]);
 
       if (error) throw error;
 
@@ -22,12 +23,12 @@ const roleController = {
   // Mengedit role berdasarkan role_id
   editRole: async (request, h) => {
     const { role_id } = request.params;
-    const { roleWorker } = request.payload;
+    const { roleworker } = request.payload;
 
     try {
       const { data, error } = await supabase
         .from('role_worker')
-        .update({ roleWorker, updated_at: new Date() })
+        .update({ roleworker })
         .eq('role_id', role_id)
         .select();
 
@@ -64,6 +65,13 @@ const roleController = {
     const { role_id } = request.params;
 
     try {
+      const { error: updateError} = await supabase
+        .from("workers")
+        .update({ role_id: 1 })
+        .eq("role_id", role_id);
+
+      if (updateError) throw updateError;
+
       const { data, error } = await supabase
         .from('role_worker')
         .delete()
